refactor: storageEffect (#36)
