+++ src/shared/persist/index.ts
@@ -1,2 +1,1 @@
 export * from './effect';
-export * from './store';
